refactor: CRTableListColumnList+SchemaDetails
