+++ static_report/src/components/shared/NoData.tsx
@@ -1,0 +1,9 @@
+import { Text } from '@chakra-ui/react';
+
+export function NoData() {
+  return (
+    <Text mt={7} color="gray.500">
+      No data available
+    </Text>
+  );
+}
